Moved bills tab to a new /licences module as requested in PR
