+++ src/internal/modules/billing/routes.js
@@ -1,11 +1,9 @@
 const billRunRoutes = require('./routes/bill-run');
 const twoPartTariff = require('./routes/two-part-tariff');
 const createBillRun = require('./routes/create-bill-run');
-const view = require('./routes/view');
 
 module.exports = [
   ...Object.values(billRunRoutes),
   ...Object.values(twoPartTariff),
-  ...Object.values(createBillRun),
-  ...Object.values(view)
+  ...Object.values(createBillRun)
 ];
