+++ src/internal/modules/licences/controllers/bills-tab.js
