+++ src/internal/modules/licences/routes.js
@@ -1,0 +1,5 @@
+const billsTab = require('./routes/bills-tab');
+
+module.exports = [
+  ...Object.values(billsTab)
+];
