+++ src/internal/modules/licences/routes/bills-tab.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Joi = require('@hapi/joi');
-const controller = require('../controllers/view');
+const controller = require('../controllers/bills-tab');
 const { scope } = require('../../../lib/constants');
 
 const allowedScopes = [scope.billing];
