+++ src/internal/views/nunjucks/billing/macros/invoices-table.njk
