+++ src/shared/view/nunjucks/macros/google-analytics.njk
@@ -4,7 +4,7 @@
 
   nonce: This should be passed as the value to attach to the script tag
 
-  tracking: An object set up in the view.js file containing:
+  tracking: An object set up in the bills-tab.js file containing:
     propertyId: The unique google analytics identifier
     isLoggedIn: Is the user logged in
     lastLogin: The last log in date for the user
